test: split up tests
